refactor: move social-media to data folder
