+++ src/infra/db/prisma/data/social-media/find-many-social-medias/find-many-social-medias-prisma-repo.ts
@@ -1,6 +1,6 @@
 import { type SocialMediaModel } from '@/domain/models/social-media/social-media-model'
 import { type FindManySocialMediasRepo } from '@/usecases/contracts/db/social-media'
-import { PrismaHelper } from '../../helpers/prisma-helper'
+import { PrismaHelper } from '@/infra/db/prisma/helpers/prisma-helper'
 
 export class FindManySocialMediasPrismaRepo implements FindManySocialMediasRepo {
   async execute (): Promise<SocialMediaModel[] | []> {
