+++ src/infra/db/prisma/data/social-media/find-social-media-by-id/find-social-media-by-id-prisma-repo.ts
@@ -1,6 +1,6 @@
 import { type SocialMediaModel } from '@/domain/models/social-media/social-media-model'
 import { type FindSocialMediaByIdRepo } from '@/usecases/contracts/db/social-media/find-social-media-by-id-repo'
-import { PrismaHelper } from '../../helpers/prisma-helper'
+import { PrismaHelper } from '@/infra/db/prisma/helpers/prisma-helper'
 
 export class FindSocialMediaByIdPrismaRepo implements FindSocialMediaByIdRepo {
   async execute (socialMediaId: string): Promise<null | SocialMediaModel> {
