+++ src/infra/db/prisma/data/social-media/find-social-media-by-name/find-social-media-by-name-prisma-repo.ts
@@ -1,6 +1,6 @@
 import { type SocialMediaModel } from '@/domain/models/social-media/social-media-model'
 import { type FindSocialMediaByNameRepo } from '@/usecases/contracts/db/social-media/find-social-media-by-name-repo'
-import { PrismaHelper } from '../../helpers/prisma-helper'
+import { PrismaHelper } from '@/infra/db/prisma/helpers/prisma-helper'
 
 export class FindSocialMediaByNamePrismaRepo implements FindSocialMediaByNameRepo {
   async execute (name: string): Promise<SocialMediaModel | null> {
