+++ src/main/factories/infra/db/prisma/social-media/add-social-media-prisma-repo-factory.ts
@@ -1,4 +1,4 @@
-import { AddSocialMediaPrismaRepo } from '@/infra/db/prisma/social-media/add-social-media/add-social-media-prisma-repo'
+import { AddSocialMediaPrismaRepo } from '@/infra/db/prisma/data/social-media/add-social-media/add-social-media-prisma-repo'
 import { type AddSocialMediaRepo } from '@/usecases/contracts/db/social-media'
 
 export const makeAddSocialMediaPrismaRepo = (): AddSocialMediaRepo => {
