+++ src/main/factories/infra/db/prisma/social-media/find-many-social-medias-prisma-repo-factory.ts
@@ -1,4 +1,4 @@
-import { FindManySocialMediasPrismaRepo } from '@/infra/db/prisma/social-media/find-many-social-medias/find-many-social-medias-prisma-repo'
+import { FindManySocialMediasPrismaRepo } from '@/infra/db/prisma/data/social-media/find-many-social-medias/find-many-social-medias-prisma-repo'
 
 export const makeFindManySocialMediasPrismaRepo = (): FindManySocialMediasPrismaRepo => {
   return new FindManySocialMediasPrismaRepo()
