+++ src/main/factories/infra/db/prisma/social-media/find-social-media-by-id-repo-factory.ts
@@ -1,4 +1,4 @@
-import { FindSocialMediaByIdPrismaRepo } from '@/infra/db/prisma/social-media/find-social-media-by-id/find-social-media-by-id-prisma-repo'
+import { FindSocialMediaByIdPrismaRepo } from '@/infra/db/prisma/data/social-media/find-social-media-by-id/find-social-media-by-id-prisma-repo'
 import { type FindSocialMediaByIdRepo } from '@/usecases/contracts/db/social-media/find-social-media-by-id-repo'
 
 export const makeFindSocialMediaByIdRepo = (): FindSocialMediaByIdRepo => {
